v1.0.3 - Improved Tests, High Score, and UI Tweak
Added better unit and functional testing for the basic game
functionality. Coverage is not 100% due to functions specified
with defaults which are never replaced in the code, so this is
acceptable, if slightly smelly.

Added a high score counter to track the session's current
longest streak of correct answers.

Slight UI tweak to the flash element, to display a red
background on the element when the user has answered incorrectly.