v1.0.6 - Dynamic Version, Displayed in Page
Moved version to a dynamic field in pyproject.toml, taking
advantage of flit's settings to gather the version number
dynamically from __init__.py instead. This also allows me to
expose this value to the index page, so I can display the current
version on the page next to a link to the repo.

Also copied over the relevant files for Font Awesome 4.7.0, in
order to display the github icon for the repo link.